+++ kadai1/karasuex54/go.mod
@@ -1,0 +1,3 @@
+module github.com/karasuex54/gopherdojo-studyroom/kadai1
+
+go 1.17
